RandomChanges

+++ template.cpp
@@ -10,12 +10,10 @@ typedef pair<ll, ll> pl;
 #define ff first
 #define ss second
 #define pb push_back
-#define sz(v) ((ll)v.size())
+#define sz(v) ((int)v.size())
 #define all(v) v.begin(), v.end()
 
-void solve() {
-  
-}
+void solve() {}
 
 int main() {
   cin.tie(0)->sync_with_stdio(0);
